stage 1: extract fibers.runtime and make op depend on it
